requirements.txt
flake8==3.9.2
isort==5.9.1
mccabe==0.6.1
numpy==1.20.3
pandas==1.2.4
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
xlrd==2.0.1
